Share operation state across primary flows
