+++ ShuffleTask.Presentation.Tests/ShuffleTask.Presentation.Tests.csproj
@@ -10,17 +10,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="../ShuffleTask.Presentation/Models/OperationState.cs" Link="Models/OperationState.cs" />
-    <Compile Include="../ShuffleTask.Presentation/ViewModels/TasksViewModel.cs" Link="ViewModels/TasksViewModel.cs" />
-    <Compile Include="../ShuffleTask.Presentation/Services/ShuffleCoordinatorService.cs" Link="Services/ShuffleCoordinatorService.cs" />
-    <Compile Include="../ShuffleTask.Presentation/Services/IPersistentBackgroundService.cs" Link="Services/IPersistentBackgroundService.cs" />
-    <Compile Include="../ShuffleTask.Presentation/Utilities/PersistedTimerState.cs" Link="Utilities/PersistedTimerState.cs" />
-    <Compile Include="../ShuffleTask.Presentation/Utilities/PersistedSchedulerState.cs" Link="Utilities/PersistedSchedulerState.cs" />
-    <Compile Include="../ShuffleTask.Presentation/Utilities/TaskTimerSettings.cs" Link="Utilities/TaskTimerSettings.cs" />
-    <Compile Include="../ShuffleTask.Presentation/PreferenceKeys.cs" Link="PreferenceKeys.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.4.0" />
     <PackageReference Include="NSubstitute" Version="5.3.0" />
     <PackageReference Include="NUnit" Version="4.5.1" />
@@ -33,6 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\ShuffleTask.Presentation.Shared\ShuffleTask.Presentation.Shared.csproj" />
     <ProjectReference Include="..\ShuffleTask.Application\ShuffleTask.Application.csproj" />
     <ProjectReference Include="..\ShuffleTask.Domain\ShuffleTask.Domain.csproj" />
     <ProjectReference Include="..\ShuffleTask.Tests\ShuffleTask.Tests.csproj" />
