+++ ShuffleTask.Presentation/MauiProgram.cs
@@ -157,13 +157,12 @@ public static partial class MauiProgram
             return transport;
         });
         builder.Services.AddSingleton<NetworkedEventAggregator>();
-        builder.Services.AddSingleton<INetworkSyncService, NetworkSyncService>();
+        builder.Services.AddSingleton<ISyncExchangeService, SyncExchangeService>();
+        builder.Services.AddSingleton<NetworkSyncService>();
+        builder.Services.AddSingleton<INetworkSyncService>(sp => sp.GetRequiredService<NetworkSyncService>());
 
         builder.Services.AddSingleton<TaskStartedAsyncHandler>();
         builder.Services.AddSingleton<TimeUpNotificationAsyncHandler>();
-        builder.Services.AddSingleton<TaskManifestAnnouncedAsyncHandler>();
-        builder.Services.AddSingleton<TaskManifestRequestAsyncHandler>();
-        builder.Services.AddSingleton<TaskBatchResponseAsyncHandler>();
     }
 
     private static void InitNetworkSync()
@@ -176,9 +175,6 @@ public static partial class MauiProgram
         {
             aggregator.SubscribeToEventType(_services!.GetRequiredService<TaskStartedAsyncHandler>());
             aggregator.SubscribeToEventType(_services!.GetRequiredService<TimeUpNotificationAsyncHandler>());
-            aggregator.SubscribeToEventType(_services!.GetRequiredService<TaskManifestAnnouncedAsyncHandler>());
-            aggregator.SubscribeToEventType(_services!.GetRequiredService<TaskManifestRequestAsyncHandler>());
-            aggregator.SubscribeToEventType(_services!.GetRequiredService<TaskBatchResponseAsyncHandler>());
         }, TaskScheduler.Default);
     }
