+++ ShuffleTask.Presentation/Views/EditTaskPage.xaml
@@ -2,7 +2,7 @@
 <?xaml-comp compile="true"?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:vm="clr-namespace:ShuffleTask.ViewModels"
+             xmlns:vm="clr-namespace:ShuffleTask.ViewModels;assembly=ShuffleTask.Presentation.Shared"
              xmlns:controls="clr-namespace:ShuffleTask.Controls"
              x:Class="ShuffleTask.Views.EditTaskPage"
              x:DataType="vm:EditTaskViewModel"
